refactor(docs): sort parents alphabetically in table and add unit tests
